+++ savant_deepstream/nvtracker/src/batching_operator/submit.rs
@@ -4,7 +4,7 @@ use deepstream_buffers::{BatchState, SavantIdMetaKind};
 use parking_lot::Mutex;
 use savant_core::primitives::frame::VideoFrameProxy;
 use std::collections::HashMap;
-use std::sync::atomic::{AtomicBool, AtomicI64, Ordering};
+use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::sync::Arc;
 
 use super::config::NvTrackerBatchingOperatorConfig;
@@ -19,7 +19,7 @@ pub(super) struct SubmitContext {
     pub(super) batch_formation: TrackerBatchFormationCallback,
     pub(super) state: Arc<Mutex<BatchState<FramePair>>>,
     pub(super) pending_batches: PendingMap,
-    pub(super) next_batch_id: Arc<AtomicI64>,
+    pub(super) next_batch_id: Arc<AtomicU64>,
     pub(super) source_frame_counters: Arc<Mutex<HashMap<String, u32>>>,
     pub(super) nvtracker: Arc<Mutex<NvTracker>>,
     pub(super) shutdown_flag: Arc<AtomicBool>,
@@ -53,7 +53,7 @@ impl SubmitContext {
             )));
         }
 
-        let batch_id = self.next_batch_id.fetch_add(1, Ordering::Relaxed);
+        let batch_id = self.next_batch_id.fetch_add(1, Ordering::Relaxed) as u128;
         ids.insert(0, SavantIdMetaKind::Batch(batch_id));
 
         let (tracked_frames, frame_nums, per_source_increments) = {
