+++ savant_deepstream/nvtracker/tests/test_nvsort_resolution_reset.rs
@@ -69,7 +69,7 @@ fn make_frame(source: &str, rois: Vec<Roi>, w: u32, h: u32, pts_ns: Option<u64>)
 
 fn frame_ids(n: usize) -> Vec<SavantIdMetaKind> {
     (0..n)
-        .map(|i| SavantIdMetaKind::Frame(i as i64 + 1))
+        .map(|i| SavantIdMetaKind::Frame(i as u128 + 1))
         .collect()
 }
